refactor(#3992): extract the shared String-proto ToString into its own module
The LOC-regrowth ratchet (#3102) flagged +99 lines on
`src/codegen/array-object-proto.ts`. Extracting rather than granting a
`loc-budget-allow:` is the better answer here, because the growth was itself
duplication: the ToPrimitive-first `? ToString(x)` sequence had been inlined
at four call sites with an if/else fallback each.

`src/codegen/string-proto-tostring.ts` now owns `emitStringProtoToStringFlat`
and `NO_ARG_STRING_MEMBER_HELPER`, the same way `char-at-transfer.ts` and
`string-proto-substring.ts` own their lowerings — the dispatcher stays a
dispatcher. Each of the four sites is now a single call.

`buildTransferredNativeProtoCallInstrs` takes its four local/global indices as
one `slots` object instead of four positional params, which also removes the
six-line call formatting at the one call site.

Both touched god-files now sit BELOW their budgets, not merely at them:
  array-object-proto.ts  2620 -> 2519  (budget 2521)
  closure-exports.ts     1678 -> 1669  (budget 1675)

Behaviour re-verified after the move, not assumed: tsc clean, biome clean,
prettier clean, stack-balance gate OK, and the 10 #3992 tests still pass.


✓
