+++ src/codegen/string-proto-tostring.ts
@@ -1,0 +1,72 @@
+// Copyright (c) 2026 Loopdive GmbH. Licensed under Apache-2.0 WITH LLVM-exception.
+/**
+ * (#3992) The shared `? ToString(x)` lowering for reflective
+ * `String.prototype.<member>` closure bodies, plus the table of no-argument
+ * string-returning members that share one body.
+ *
+ * Both live here rather than in the `array-object-proto` dispatcher for the
+ * same reason `char-at-transfer.ts` and `string-proto-substring.ts` do: the
+ * dispatcher should stay a dispatcher, and this is the one rule those bodies
+ * kept re-deriving.
+ */
+import type { Instr } from "../ir/types.js";
+import { runtimeToPrimitiveInstrs } from "./coercion-engine.js";
+import type { CodegenContext, FunctionContext } from "./context/types.js";
+
+/**
+ * The reflective `String.prototype` members that take NO arguments and return a
+ * STRING, mapped to the standalone-native helper implementing them — the SAME
+ * helpers the direct `"x".trim()` / `"x".toLowerCase()` path uses in
+ * `string-ops.ts`.
+ *
+ * (#1470) `toLocale{Lower,Upper}Case` fold onto the non-locale helpers: absent
+ * ECMA-402, §22.1.3.{27,29} are defined to match `to{Lower,Upper}Case` except
+ * for locale-sensitive mappings, and a standalone module carries no ICU tables.
+ * The direct path already makes exactly this substitution, so routing the
+ * reflective body to the same helper keeps the two paths in agreement rather
+ * than inventing a second answer.
+ */
+export const NO_ARG_STRING_MEMBER_HELPER: Readonly<Record<string, string>> = {
+  trim: "__str_trim",
+  trimStart: "__str_trimStart",
+  trimEnd: "__str_trimEnd",
+  toLowerCase: "__str_toLowerCase",
+  toUpperCase: "__str_toUpperCase",
+  toLocaleLowerCase: "__str_toLowerCase",
+  toLocaleUpperCase: "__str_toUpperCase",
+};
+
+/**
+ * Emit `flatten(? ToString(param<paramIdx>))`, leaving the flat `$NativeString`
+ * on the stack.
+ *
+ * §22.1.3 says `ToString(thisValue)`, and ToString of an OBJECT is
+ * `ToPrimitive(input, string)` first — that is what consults `toString` /
+ * `valueOf` (§7.1.1 OrdinaryToPrimitive). The generic reflective bodies used to
+ * skip straight to `$__any_to_string`, which stringifies an object receiver
+ * structurally: `new Object(true)` came out as `"[object Object]"`, and a user
+ * `toString` was never called at all.
+ *
+ * That was invisible until #3992 fixed the transferred-method dispatch, because
+ * every such call previously answered `null` before reaching a body. The two
+ * members that DID work (`charAt`, `substring`) are exactly the two whose
+ * bespoke bodies already did the ToPrimitive step — so this was the third and
+ * fourth spelling of one rule.
+ *
+ * When `__to_primitive` is unavailable (host/gc lowering) the previous
+ * byte-identical sequence is emitted instead, so non-standalone output is
+ * unchanged.
+ */
+export function emitStringProtoToStringFlat(
+  ctx: CodegenContext,
+  fctx: FunctionContext,
+  paramIdx: number,
+  anyToStrIdx: number,
+  flattenIdx: number,
+): void {
+  const toPrimitive = runtimeToPrimitiveInstrs(ctx, "string");
+  const body: Instr[] = [{ op: "local.get", index: paramIdx }];
+  if (toPrimitive !== null) body.push(...toPrimitive);
+  body.push({ op: "any.convert_extern" }, { op: "call", funcIdx: anyToStrIdx }, { op: "call", funcIdx: flattenIdx });
+  for (const instr of body) fctx.body.push(instr);
+}
